requirements.txt
Flask==0.10.1
flask-mongoengine==0.7.0
Flask-Bcrypt
flask-login==0.2.7
gevent-socketio
